requirements.txt
asgiref==3.5.2
Babel==2.10.1
Django==4.0.5
django-bootstrap-datepicker-plus==4.0.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-money==2.1.1
gunicorn==20.1.0
Pillow==9.1.1
py-moneyed==1.2
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
